% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynasimR-package.R
\docType{package}
\name{dynasimR-package}
\alias{dynasimR-package}
\alias{dynasimR}
\title{dynasimR: Dynamic Agent-Node Simulation Analysis}
\description{
Analysis and visualisation layer for discrete-event, agent-based and
node-actor simulation outputs. Primary application is the
MEDTACS-SIM military medical simulation (chain of resuscitation,
doctrine effects, autonomy trade-offs) and the REHASIM rehabilitation
flow simulation (FIM trajectories, waiting-gap index, supply-demand).
}
\details{
The package is schema-harmonised so that both simulations can be
analysed with the same API. See the vignettes for worked examples.
}
\section{Key functions}{

\describe{
\item{Data I/O}{\code{\link[=read_simulation]{read_simulation()}}, \code{\link[=load_example_data]{load_example_data()}},
\code{\link[=validate_dynasimR_data]{validate_dynasimR_data()}}}
\item{Survival}{\code{\link[=km_estimate]{km_estimate()}}, \code{\link[=cox_model]{cox_model()}}}
\item{Doctrine}{\code{\link[=doctrine_effect]{doctrine_effect()}}}
\item{Autonomy}{\code{\link[=al_efficiency]{al_efficiency()}}}
\item{IHL}{\code{\link[=compute_ihl_index]{compute_ihl_index()}}}
\item{Patient flow}{\code{\link[=role_throughput]{role_throughput()}}, \code{\link[=detect_bottlenecks]{detect_bottlenecks()}}}
\item{UAV}{\code{\link[=uav_comparison]{uav_comparison()}}}
\item{Sensitivity}{\code{\link[=morris_screening]{morris_screening()}}, \code{\link[=tornado_data]{tornado_data()}}}
\item{REHASIM}{\code{\link[=fim_trajectory_analysis]{fim_trajectory_analysis()}},
\code{\link[=compute_waiting_gap_index]{compute_waiting_gap_index()}}, \code{\link[=mil_civil_effect]{mil_civil_effect()}},
\code{\link[=spatial_supply_demand]{spatial_supply_demand()}}, \code{\link[=compute_rtd_analysis]{compute_rtd_analysis()}}}
\item{Plots}{\code{\link[=plot_km]{plot_km()}}, \code{\link[=plot_forest]{plot_forest()}}, \code{\link[=plot_al_tradeoff]{plot_al_tradeoff()}},
\code{\link[=plot_scenario_heatmap]{plot_scenario_heatmap()}}, \code{\link[=plot_doctrine]{plot_doctrine()}}, \code{\link[=plot_timeline]{plot_timeline()}},
\code{\link[=plot_map]{plot_map()}}, \code{\link[=plot_fim_curves]{plot_fim_curves()}}, \code{\link[=plot_sdi_map]{plot_sdi_map()}},
\code{\link[=plot_cost_effectiveness]{plot_cost_effectiveness()}}}
\item{Export}{\code{\link[=export_figure]{export_figure()}}, \code{\link[=export_latex_table]{export_latex_table()}},
\code{\link[=fill_placeholders]{fill_placeholders()}}}
\item{Shiny}{\code{\link[=launch_app]{launch_app()}}, \code{\link[=check_app_dependencies]{check_app_dependencies()}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rabanheller/dynasimR}
  \item Report bugs at \url{https://github.com/rabanheller/dynasimR/issues}
}

}
\author{
\strong{Maintainer}: Raban A. Heller \email{rabanheller@bundeswehr.org}

}
\keyword{internal}
