% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fim_curves.R
\name{plot_fim_curves}
\alias{plot_fim_curves}
\title{FIM trajectory plot (REHASIM)}
\usage{
plot_fim_curves(fim_result, show_ci = TRUE)
}
\arguments{
\item{fim_result}{A \code{dynasimR_fim} object.}

\item{show_ci}{Logical. Show 95\% CI ribbon. Default \code{TRUE}.}
}
\value{
A ggplot2 object.
}
\description{
Plots longitudinal FIM trajectories from the output of
\code{\link[=fim_trajectory_analysis]{fim_trajectory_analysis()}} (with \code{longitudinal = TRUE}).
}
