% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest plot of Cox hazard ratios}
\usage{
plot_forest(cox_result, reference_label = "No effect (HR = 1)", ...)
}
\arguments{
\item{cox_result}{A \code{dynasimR_cox} object.}

\item{reference_label}{Character. Label for HR = 1 reference line.
Default \code{"No effect (HR = 1)"}.}

\item{...}{Unused.}
}
\value{
A ggplot2 object.
}
\description{
Forest plot of Cox hazard ratios
}
