% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_color_group_dynasimR}
\alias{scale_color_group_dynasimR}
\title{Colour scale for entity group}
\usage{
scale_color_group_dynasimR(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Applies dynasimR group colours (\code{A}/\code{B}/\code{C}) to a ggplot2
\code{color} aesthetic.
}
