% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_color_policy_dynasimR}
\alias{scale_color_policy_dynasimR}
\title{Colour scale for policy}
\usage{
scale_color_policy_dynasimR(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Colour scale for policy
}
