% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_group_dynasimR}
\alias{scale_fill_group_dynasimR}
\title{Fill scale for entity group}
\usage{
scale_fill_group_dynasimR(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
As \code{\link[=scale_color_group_dynasimR]{scale_color_group_dynasimR()}} but for the \code{fill} aesthetic.
}
