% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_identity_dynasimR}
\alias{scale_fill_identity_dynasimR}
\title{Fill scale for combatant identity}
\usage{
scale_fill_identity_dynasimR(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
As \code{\link[=scale_color_identity_dynasimR]{scale_color_identity_dynasimR()}} but for the \code{fill} aesthetic.
}
