% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{cox_model}
\alias{cox_model}
\title{Cox proportional-hazards model for simulation data}
\usage{
cox_model(
  data,
  endpoint = c("stage2", "overall", "service"),
  covariates = c("scenario", "severity", "group"),
  reference_scenario = "A-S00"
)
}
\arguments{
\item{data}{A \code{dynasimR_data} object or entity tibble.}

\item{endpoint}{See \code{\link[=km_estimate]{km_estimate()}}.}

\item{covariates}{Character vector. Covariates entering the Cox
model. Default \code{c("scenario","severity","group")}.}

\item{reference_scenario}{Character. Reference level for the
\code{scenario} factor. Default \code{"A-S00"}. Use \code{"B-S00"} for Profile B.}
}
\value{
An S3 object of class \code{dynasimR_cox} with slots \code{fit},
\code{tidy}, \code{ph_test}, \code{forest_data} and \code{params}.
}
\description{
Cox proportional-hazards model for simulation data
}
